chore(frontend): rename folder
